Minor Refactors
