+++ Inoculated.Test/Interceptor.cs
@@ -4,12 +4,12 @@ using Inoculator.Builder;
 
 public class LogEntrencyAttribute : InterceptorAttribute
 {
-    public override void OnEntry(Metadata method)
+    public override void OnEntry(MethodData method)
     {
         Console.WriteLine($"Started Method {method.Name}");
     }
 
-    public override void OnExit(Metadata method)
+    public override void OnExit(MethodData method)
     {
         Console.WriteLine($"Finished Method {method.Name} with {method.ReturnValue}");
     }
@@ -17,12 +17,12 @@ public class LogEntrencyAttribute : InterceptorAttribute
 
 public class ElapsedTimeAttribute : InterceptorAttribute
 {
-    public override void OnEntry(Metadata method)
+    public override void OnEntry(MethodData method)
     {
         method.EmbededResource = Stopwatch.StartNew();        
     }
 
-    public override void OnExit(Metadata method)
+    public override void OnExit(MethodData method)
     {
         var sw = (Stopwatch)method.EmbededResource;
         sw.Stop();
